+++ jbg1/src/main/java/com/example/likelion_2/member/controller/MemberController.java
@@ -1,15 +1,13 @@
 package com.example.likelion_2.member.controller;
 
 import com.example.likelion_2.member.domain.Member;
-import com.example.likelion_2.member.repository.MemberRepository;
+import com.example.likelion_2.member.service.MemberService;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,7 +15,7 @@ import java.util.List;
 @Controller
 @RequiredArgsConstructor
 public class MemberController {
-    private final MemberRepository memberRepository;
+    private final MemberService memberService;
 
     @GetMapping("/member")
     public String signupPage() {
@@ -30,13 +28,13 @@ public class MemberController {
         int age = Integer.parseInt(request.getParameter("age"));
         String password = request.getParameter("password");
         Member member = new Member(name, age, password);
-        memberRepository.save(member);
+        memberService.addMember(member);
         return "redirect:/";
     }
 
     @GetMapping("/member/list")
     public String listPage(Model model) {
-        List<Member> members = new ArrayList<>(memberRepository.getStore().values());
+        List<Member> members = new ArrayList<>(memberService.getMemberCollection());
         model.addAttribute("members", members);
         return "list";
     }
